// src/types/mainGame.ts
export interface BaseballGameData {
  AWAY_ID: string;
  AWAY_NM: string;
  B1_BAT_ORDER_NO: number;
  B2_BAT_ORDER_NO: number;
  B3_BAT_ORDER_NO: number;
  BALL_CN: number;
  B_D_PIT_P_ID: number | null;
  B_D_PIT_P_NM: string;
  B_GROUP_SC: number | null;
  B_PIT_P_ID: number;
  B_PIT_P_NM: string;
  B_P_ID: number;
  B_P_NM: string;
  B_RANK_NO: number;
  B_SCORE_CN: string;
  CANCEL_SC_ID: string;
  CANCEL_SC_NM: string;
  CHECK_SWING_CK: boolean;
  DETAIL_SC: null;
  GAME_INN_NO: number;
  GAME_NO: number | null;
  GAME_RESULT_CK: number;
  GAME_SC_ID: number;
  GAME_SC_NM: string;
  GAME_STATE_SC: string;
  GAME_TB_SC: string;
  GAME_TB_SC_NM: string;
  G_DT: string;
  G_DT_TXT: string;
  G_ID: string;
  G_TM: string;
  HEADER_NO: number;
  HOME_ID: string;
  HOME_NM: string;
  IE_ENTRY_CK: number;
  KBOT_SE: number;
  LE_ID: number;
  LINEUP_CK: number;
  L_PIT_P_ID: number | null;
  L_PIT_P_NM: string;
  OUT_CN: number;
  ROUND_SC: null;
  SCORE_CK: string;
  SEASON_ID: number;
  SR_ID: number;
  START_PIT_CK: number;
  STRIKE_CN: number;
  SV_PIT_P_ID: number | null;
  SV_PIT_P_NM: string;
  S_NM: string;
  TV_IF: string;
  T_D_PIT_P_ID: number | null;
  T_D_PIT_P_NM: string;
  T_GROUP_SC: number | null;
  T_PIT_P_ID: number;
  T_PIT_P_NM: string;
  T_P_ID: number;
  T_P_NM: string;
  T_RANK_NO: number;
  T_SCORE_CN: string;
  VOD_CK: number;
  VS_GAME_CN: number;
  W_PIT_P_ID: number | null;
  W_PIT_P_NM: string;
}

export interface RankingData {
  Class: null;
  Id: null;
  OnClick: null;
  Style: null;
  Value: null;
  row: {
    Class: null;
    ColSpan: null;
    RowSpan: null;
    Scope: null;
    Text: string;
    TypeObj: string;
    Width: null;
  }[];
}